Refactor Layers
