+++ src/app/data/services/user.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {User} from "../models/user";
+import {User} from "../../domain/models/user";
 
 @Injectable({
   providedIn: 'root'
