+++ src/app/presentation/components/oink-content/oink-content.component.spec.ts
