+++ src/app/presentation/pages/account-data-sign-up/account-data-sign-up.page.scss
