+++ src/app/presentation/pages/auth/auth.page.html
@@ -22,7 +22,7 @@
   <ion-grid>
     <ion-row class="center-row" id="one">
       <ion-col class="ion-text-center" size="12">
-        <img alt="Coink APP" height="155" ngSrc="../../../assets/images/coinkPicture.svg" width="126">
+        <img alt="Coink APP" height="155" ngSrc="assets/images/coinkPicture.svg" width="126">
       </ion-col>
       <ion-col class="ion-text-center" size="12" style="margin-top: 40px">
         <ion-text>
