+++ src/app/presentation/pages/phone-sign-up/phone-sign-up.page.html
@@ -69,7 +69,7 @@
       </ion-row>
       <ion-row>
         <ion-col>
-          <img (click)="delete()" alt="Backspace" src="../../../assets/images/backspace.svg">
+          <img (click)="delete()" alt="Backspace" src="assets/images/backspace.svg">
         </ion-col>
         <ion-col>
           <ion-text (click)="write(0)">
