+++ src/app/presentation/pages/phone-sign-up/phone-sign-up.page.scss
