+++ src/app/presentation/pages/terms-sign-up/components/modal/modal.component.html
