+++ src/app/presentation/pages/terms-sign-up/terms-sign-up.page.scss
