+++ src/main.ts
@@ -1,6 +1,6 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import {AppModule} from './app/presentation/app.module';
 
 platformBrowserDynamic().bootstrapModule(AppModule)
   .catch(err => console.log(err));
